+++ src/assets/langs/es.json
@@ -144,7 +144,7 @@
   },
   "contact": {
     "content": {
-      "main": {
+      "info": {
         "title": "Contacto",
         "description": "Estás a un paso de olvidarte de las multas y cierres. Escríbenos y nos pondremos en contacto para más detalles."
       }
