+++ src/components/sections/services/services.scss
@@ -6,7 +6,8 @@
   padding: 10dvh 0;
   box-sizing: border-box;
 
-  background: url("../../../assets/images/bg-services.jpg") center/cover fixed;
+  background: url("../../../assets/images/services/bg-services.webp")
+    center/cover fixed;
 
   display: flex;
   justify-content: center;
