+++ src/components/sections/welcome/welcome.scss
@@ -6,7 +6,7 @@
   min-height: 100vh !important;
   box-sizing: border-box;
 
-  background: url("../../../assets/images/bg-welcome.jpg") center/cover fixed;
+  background: url("../../../assets/images/welcome/bg-welcome.webp") center/cover fixed;
 
   display: flex;
   justify-content: center;
