+++ src/types.d.ts
@@ -3,17 +3,7 @@ declare module "*.svg" {
   export default content;
 }
 
-declare module "*.png" {
+declare module "*.webp" {
   const content: string;
   export default content;
 }
-
-declare module "*.jpg" {
-  const content: string;
-  export default content;
-}
-
-declare module "*.jpeg" {
-  const content: string;
-  export default content;
-}
